+++ components/tr.lua
@@ -363,3 +363,20 @@ function SAO:gameSpellAlertsChangeLater(optionSequence)
     };
     return string.format(tr(gameSpellAlertsChangeLaterTranslations), optionSequence);
 end
+
+-- Translate "Keep the game's native spell alerts"
+function SAO:doNotAskDisableGameAlerts()
+    local doNotAskDisableGameAlertsTranslations = {
+        ["en"] = "Keep the game's native spell alerts",
+        ["de"] = "Die nativen Zauberwarnungen des Spiels beibehalten",
+        ["fr"] = "Conserver les alertes de sort natives du jeu",
+        ["es"] = "Mantener las alertas de hechizo nativas del juego",
+        ["ru"] = "Сохранить родные оповещения о заклинаниях игры",
+        ["it"] = "Mantenere gli avvisi sugli incantesimi nativi del gioco",
+        ["pt"] = "Manter os alertas de feitiço nativos do jogo",
+        ["ko"] = "게임의 기본 주문 경고 유지",
+        ["zh"] = "保留游戏的原生法术警报",
+        ["zhTW"] = "保留遊戲的原生法術警報",
+    };
+    return tr(doNotAskDisableGameAlertsTranslations);
+end